+++ src/frontend/pages/index.tsx
@@ -6,19 +6,22 @@ import * as S from '../styles/Home.styled';
 import { useQuery } from 'react-query';
 import ApiGateway from '../gateways/Api.gateway';
 import Banner from '../components/Banner';
+import { CypressFields } from '../utils/Cypress';
 
 const Home: NextPage = () => {
   const { data: productList = [] } = useQuery('products', ApiGateway.listProducts);
 
   return (
     <Layout>
-      <S.Home>
+      <S.Home data-cy={CypressFields.HomePage}>
         <Banner />
         <S.Container>
           <S.Row>
             <S.Content>
               <S.HotProducts>
-                <S.HotProductsTitle id="hot-products">Hot Products</S.HotProductsTitle>
+                <S.HotProductsTitle data-cy={CypressFields.HotProducts} id="hot-products">
+                  Hot Products
+                </S.HotProductsTitle>
                 <ProductList productList={productList} />
               </S.HotProducts>
             </S.Content>