+++ CHANGELOG.md
@@ -65,5 +65,7 @@ significant modifications will be credited to OpenTelemetry Authors.
 ([#273](https://github.com/open-telemetry/opentelemetry-demo/pull/273))
 * Reimplemented Frontend app using [Next.js](https://nextjs.org/) Browser client
 ([#236](https://github.com/open-telemetry/opentelemetry-demo/pull/236))
+* Added Frontend [Cypress](https://www.cypress.io/) E2E tests
+([#298](https://github.com/open-telemetry/opentelemetry-demo/pull/298))
 * Add HEALTHCHECK to all services
 ([#296](https://github.com/open-telemetry/opentelemetry-demo/pull/296))